clean up link tests
