+++ src/Halcyon.Tests/Halcyon.Tests.csproj
@@ -82,6 +82,7 @@
   <ItemGroup>
     <Compile Include="GlobalSuppressions.cs" />
     <Compile Include="HAL\ControllerExtensionsTests.cs" />
+    <Compile Include="HAL\EmbeddAttributeTests.cs" />
     <Compile Include="HAL\HALModelConfigTests.cs" />
     <Compile Include="HAL\HALModelExtensionTests.cs" />
     <Compile Include="HAL\HALResponseTests.cs" />
